+++ hutool-json/src/test/java/cn/hutool/json/Issue2365Test.java
@@ -1,3 +1,5 @@
+package cn.hutool.json;
+
 import cn.hutool.json.JSONUtil;
 import lombok.Data;
 import org.junit.Assert;